requirements.txt
Flask==0.10.1
Flask-Migrate==1.1.0
Flask-Script==0.6.6
Flask-SQLAlchemy==1.0
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
SQLAlchemy==0.8.4
Werkzeug==0.9.4
alembic==0.6.2
blinker==1.3
itsdangerous==0.23
gunicorn==18.0
psycopg2==2.5.1
mandrill==1.0.57
ipython==1.1.0
python-twitter==2.0
